+++ Sources/AudioKit/Taps/RawDataTap.swift
@@ -49,30 +49,6 @@ open class RawDataTap: BaseTap {
     }
 }
 
-public protocol Tap {
-    func handleTap(buffer: AVAudioPCMBuffer, at time: AVAudioTime) async
-}
-
-public func install(tap: Tap, on input: Node, bufferSize: UInt32) {
-
-    // Should we throw an exception instead?
-    guard input.avAudioNode.engine != nil else {
-        Log("The tapped node isn't attached to the engine")
-        return
-    }
-
-    let bus = 0 // Should be a ctor argument?
-    input.avAudioNode.installTap(onBus: bus,
-                                 bufferSize: bufferSize,
-                                 format: nil,
-                                 block: { (buffer, time) in
-        Task {
-            await tap.handleTap(buffer: buffer, at: time)
-        }
-    })
-
-}
-
 public actor RawDataTap2: Tap {
 
     /// Callback type
@@ -95,7 +71,7 @@ public actor RawDataTap2: Tap {
                 data.append(tail[offset + Int(idx)])
             }
         }
-        
+
         handler(data)
     }
 }
